+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # master (unreleased)
 
 Development tools:
-* adds `.ruby-version``
+* add `.ruby-version``
+* add `Makefile`
 
 # v6.1.0 (September 15, 2023)